Minor structure refactoring 🔨
