+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-Authentication/Components/SplittedComponents/LoaderComponent/index.tsx
@@ -1,5 +1,5 @@
 import React, {FunctionComponent} from 'react';
-import loader from '../../../../ignitus-Assets/ignitus-Logos/ignitusLoader.gif';
+import loader from '../../../../../ignitus-Assets/ignitus-Logos/ignitusLoader.gif';
 
 export const Loader: FunctionComponent = () => (
   <div className="container col-lg-6 col-md-4 col-sm-6 col-9 mx-auto">
