+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-Authentication/Components/SplittedComponents/WrapperComponent/index.tsx
