+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-Authentication/Components/SplittedComponents/index.ts
@@ -1,0 +1,4 @@
+export {Wrapper} from './WrapperComponent';
+export {SideWrapper} from './SideWrapperComponent';
+export {Form} from './FormComponent';
+export {Loader} from './LoaderComponent';
