--- src/ignitus-Shared/ignitus-DesignSystem/ignitus-Templates/ignitus-Authentication/Components/partials/index.ts
@@ -1,4 +1,0 @@
-export {Wrapper} from './Wrapper';
-export {Information} from './Information';
-export {Form} from './Form';
-export {Loader} from './Loader';
